refactor(Viewer): Refactor SingleView component and extract socket logic into useSocket hook
